SW-318. Moteprobe, dont use end of frame HDLC flag twice.
When processing incoming bytes to determine the start of a frame
a check is performed to see if the the last byte was a HDLC_FLAG
and if the current one is not an HDLC flag.

This works if flowcontrol is handled before the bytes are received
otherwise the handling can consider a HDLC frame end flag and
XON/XOFF bytes as the start of a new frame which is incorrect.

To avoid this discard the HDLC end of frame flag upon the reception
of a valid HDLC frame.

squash! SW-318. Remove last_rx_byte in moteprobe